src/main/java/tech/antoniosgarbi/desafioapi/service/impl/UserDetailsServiceImpl.java
package tech.antoniosgarbi.desafioapi.service.impl;

import lombok.RequiredArgsConstructor;
import org.springframework.security.core.GrantedAuthority;
import org.springframework.security.core.authority.SimpleGrantedAuthority;
import org.springframework.security.core.userdetails.UserDetails;
import org.springframework.security.core.userdetails.UserDetailsService;
import org.springframework.security.core.userdetails.UsernameNotFoundException;
import org.springframework.stereotype.Service;
import tech.antoniosgarbi.desafioapi.model.User;
import tech.antoniosgarbi.desafioapi.repository.UserRepository;

import java.util.List;
import java.util.Optional;

@Service
@RequiredArgsConstructor
public class UserDetailsServiceImpl implements UserDetailsService {

    private final UserRepository userRepository;


    @Override
    public UserDetails loadUserByUsername(String username) throws UsernameNotFoundException {
        User user = this.findModelByEmail(username);

        GrantedAuthority authority = new SimpleGrantedAuthority(String.format("ROLE_%s", user.getRole()));

        return new org.springframework.security.core.userdetails.User(user.getEmail(), user.getPassword(), List.of(authority));
    }

    private User findModelByEmail(String username) {
        final Optional<User> optional = this.userRepository.findByEmail(username);

        if(optional.isEmpty()) {
            throw new UsernameNotFoundException("Usuario não foi encontrado");
        }
        return optional.get();
    }

}
